Enable generation of Neu in Grossdruck
which is basically a slightly modified version of Neu im Sortiment with
only a subset of the items and the items moved up one hierarchy as we
only show one format.

The headings are using a dynamic binding to make sure they have the
right size in all circumstances. This is a tad ugly as not only the font
sizes should change but also all the space after and before a title
line.

Good enough for now. We might have to look at the layout problem again
later.