Massive refactoring of numbering functionality
The numbering is now handled separately. The tree is traversed first and
a mapping of path and number is created. Later in the fop generation
stage we now longer need to traverse with map-indexed but instead we can
just look up the path in the mapping of path and number. If there is a
match we'll use the number.