Remove artifact role on static content (headers and footers)
This seems to satisfy the accessibility checker. However this might be a
bug in FOP. When the artifact role is on the static content the tags in
the last page-sequence (the verso cover) do not seem to be generated
properly. It seems weird though to not mark headers and footers as
artifacts. When you do not do this then the screen reader will read it
to you.

So in the end we might have to re-enable the artifact role on headers
and footers and get FOP fixed to emit the tags properly even for the
last page-sequence.