Use accompanying-material in braille as is
Before we were not trusting it and setting our own string under some
very specific conditions. As it happens this drops some legitimate other
use cases. So use the string as is. If there are problems solve them at
the source, i.e. fix the data in the library system.

Fixes #4