Extract target audience from marc21 xml
The Marc21 XML contains the target audience in a cryptic control
field. In order to extract that properly we need to monkey patch the
xml text node extraction because that one assumes that you want the
text of a text node normalized. In this case we don't. That's why we
define `get-controlfield` which in turn uses `text-raw`.

Also update the validation to accomodate for the new :target-audience
field.