No longer artificialy tag the print-and-braille items as :braille
Instead leave their format as is (:braille or :taktilesbuch) and
handle the rendering as a :braille book in the dispatch function of
`entry-sexp` and `entry-str`.

This gets rid of a some convoluted code when trying to duplicate only
the :braille print-and-braille items. Before since we didn't have the
original format anymore we needed an additional field named
`:really-a-taktilesbuch?`. Hello, how tacky can you get? This is now
thankfully gone.