Move the hard coded producer_mapping to the db
This should enable the user to define the mapping without having to
recompile the source and redeploy the binary.

Unfortunately the mapping is used in the prismatic schema for
validation. ATM I do not know how the schema could be created at
runtime from the data in the db. If that was feasible then we could
move the mapping to the db.