Solve the invalid fop problem with empty items more elegantly
When generating the grossdruck catalog for a future date where we have
no items and no editorial and also no recommendations we get nil as
subitems. This will generate an empty toc and hence an empty fo:flow
which is not valid. The other catalogs will at least have a block for
the title of the toc and are valid under the same circumstances. This
is because the items will be an empty map.

So to get the same effect we default the subitems to an empty map if
they are nil.