Depend on a up-to-date version of parboiled-java
So markdown parsing also works in modern Java.

endophile is based on pegdown which is
unmaintained (https://github.com/sirthias/pegdown). In fact it does not work
with newer Java versions due to its dependency on an old
parboiled-java (https://mvnrepository.com/artifact/org.parboiled/parboiled-java)
So as a quick workaround we depend on a newer parboiled-java to make the whole
thing work again.

But the real fix is to move to a well maintained clojure markdown parsing
library such as nextjournal markdown (https://github.com/nextjournal/markdown)
for markdown parsing.